General bug fixes

+++ MediaLibraryServer/Interfaces/Config/IFolderService.cs
@@ -4,6 +4,6 @@ using MediaLibraryCommon.Enums;
 
 namespace MediaLibraryServer.Interfaces.Config {
     public interface IFolderService: IAbstractLibraryService<FolderLibrary, FolderLibraryData> {
-        public FolderLibrary GetFolder(FolderType folderType, int fileSizeMb = 0);
+        public FolderLibrary GetFolder(FolderType folderType, long fileSizeMb = 0);
     }
 }
